chore: refactor into one downloads file
