+++ .vscode/settings.json
@@ -8,6 +8,7 @@
     "hookform",
     "Hubspot",
     "Nextjs",
+    "oaknational",
     "pageobjects",
     "phonenumber",
     "portabletext",
