+++ src/node-lib/curriculum-api-2023/__mocks__/index.ts
@@ -19,6 +19,7 @@ import lessonDownloadsFixtures from "@/node-lib/curriculum-api-2023/fixtures/les
 import { subjectListingFixture2023 } from "@/node-lib/curriculum-api-2023/fixtures/subjectListing.fixture";
 import { pupilProgrammeListingFixture } from "@/node-lib/curriculum-api-2023/fixtures/pupilProgrammeListing.fixture";
 import { mockUrls } from "@/node-lib/curriculum-api-2023/fixtures/teachersSiteMap.fixture";
+import { type LessonDownloadsQuery } from "@/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.query";
 
 const curriculumApi: Pick<
   CurriculumApi,
@@ -28,7 +29,6 @@ const curriculumApi: Pick<
   | "lessonListing"
   | "programmeListingPage"
   | "teachersHomePage"
-  | "lessonDownloadsCanonical"
   | "pupilLessonQuery"
   | "pupilLessonListingQuery"
   | "pupilUnitListingQuery"
@@ -85,15 +85,9 @@ const curriculumApi: Pick<
   lessonOverview: jest.fn(async () => {
     return lessonOverviewFixture();
   }),
-  lessonDownloadsCanonical: jest.fn(async () => {
-    return {
-      ...lessonDownloadsFixtures(),
-      pathways: [lessonDownloadsFixtures()],
-    };
-  }),
   lessonDownloads: jest.fn(async () => {
     return lessonDownloadsFixtures();
-  }),
+  }) as jest.Mocked<LessonDownloadsQuery>,
   unitListing: jest.fn(async () => {
     return unitListingFixture();
   }),
