+++ src/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.query.ts
@@ -1,36 +1,57 @@
+import { SyntheticUnitvariantLessons } from "@oaknational/oak-curriculum-schema";
+
 import {
-  SyntheticUnitvariantLessons,
-  syntheticUnitvariantLessonsSchema,
-} from "@oaknational/oak-curriculum-schema";
+  InputMaybe,
+  Published_Mv_Synthetic_Unitvariant_Lessons_By_Keystage_6_0_0_Bool_Exp,
+} from "../../generated/sdk";
 
 import lessonDownloadsSchema, {
   downloadsAssetData,
 } from "./lessonDownloads.schema";
 import {
   constructDownloadsArray,
-  constructLessonListingObjectArray,
-  getNextLessonsInUnit,
+  constructLessonDownloads,
 } from "./downloadUtils";
+import constructCanonicalLessonDownloads from "./constructCanonicalLessonDownloads";
 
 import errorReporter from "@/common-lib/error-reporter";
 import OakError from "@/errors/OakError";
 import { Sdk } from "@/node-lib/curriculum-api-2023/sdk";
-import keysToCamelCase from "@/utils/snakeCaseConverter";
+import lessonDownloadsCanonicalSchema from "@/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloadsCanonical.schema";
 
 const lessonDownloadsQuery =
   (sdk: Sdk) =>
-  async (args: {
-    programmeSlug: string;
-    unitSlug: string;
+  async <T>(args: {
+    programmeSlug?: string;
+    unitSlug?: string;
     lessonSlug: string;
-  }) => {
-    const res = await sdk.lessonDownloads(args);
+  }): Promise<T> => {
+    const { lessonSlug, unitSlug, programmeSlug } = args;
+
+    const browseDataWhere: InputMaybe<Published_Mv_Synthetic_Unitvariant_Lessons_By_Keystage_6_0_0_Bool_Exp> =
+      {};
+
+    const canonicalLesson = !unitSlug && !programmeSlug;
+
+    if (canonicalLesson) {
+      browseDataWhere["lesson_slug"] = { _eq: lessonSlug };
+    }
+
+    if (unitSlug) {
+      browseDataWhere["unit_slug"] = { _eq: unitSlug };
+    }
+
+    if (programmeSlug) {
+      browseDataWhere["programme_slug"] = { _eq: programmeSlug };
+    }
+
+    const res = await sdk.lessonDownloads({ lessonSlug, browseDataWhere });
 
     if (
       !res.download_assets ||
-      !res.unit_lessons ||
+      !res.browse_data ||
       res.download_assets.length === 0 ||
-      res.unit_lessons.length === 0
+      res.browse_data.length === 0
     ) {
       throw new OakError({ code: "curriculum-api/not-found" });
     }
@@ -46,7 +67,7 @@ const lessonDownloadsQuery =
       });
     }
 
-    const { download_assets, unit_lessons } = res;
+    const { download_assets, browse_data } = res;
 
     const {
       has_slide_deck_asset_object,
@@ -60,19 +81,6 @@ const lessonDownloadsQuery =
       expired,
     } = downloadsAssetData.parse(download_assets[0]);
 
-    const currentLesson = unit_lessons.find(
-      (lesson) => lesson.lesson_slug === args.lessonSlug,
-    );
-
-    const copyright = keysToCamelCase(
-      currentLesson?.lesson_data.copyright_content,
-    );
-
-    const parsedCurrentLesson =
-      syntheticUnitvariantLessonsSchema.parse(currentLesson);
-
-    // parsing currentLesson returns empty copyright object?
-
     const downloadsData = {
       hasSlideDeckAssetObject: has_slide_deck_asset_object,
       hasStarterQuiz: starter_quiz ? true : false,
@@ -83,39 +91,36 @@ const lessonDownloadsQuery =
         has_worksheet_google_drive_downloadable_version,
       hasSupplementaryAssetObject: has_supplementary_asset_object,
       isLegacy: is_legacy,
-      copyrightContent: currentLesson?.lesson_data.copyright_content,
     };
 
     const downloads = constructDownloadsArray(downloadsData);
 
-    const unitLessonsArray = constructLessonListingObjectArray(
-      res.unit_lessons as SyntheticUnitvariantLessons[],
-    );
-    const nextLessons = getNextLessonsInUnit(unitLessonsArray, args.lessonSlug);
-
-    const pageData = {
+    const canonicalLessonDownloads = constructCanonicalLessonDownloads(
       downloads,
-      programmeSlug: parsedCurrentLesson.programme_slug,
-      keyStageSlug: parsedCurrentLesson.programme_fields.keystage_slug,
-      keyStageTitle: parsedCurrentLesson.programme_fields.keystage_description,
-      lessonSlug: parsedCurrentLesson.lesson_slug,
-      lessonTitle: parsedCurrentLesson.lesson_data.title,
-      subjectSlug: parsedCurrentLesson.programme_fields.subject_slug,
-      subjectTitle: parsedCurrentLesson.programme_fields.subject,
-      unitSlug: parsedCurrentLesson.unit_data.slug,
-      unitTitle: parsedCurrentLesson.unit_data.title,
-      lessonCohort: parsedCurrentLesson.lesson_data._cohort,
-      expired: expired ? expired : null,
-      updatedAt: parsedCurrentLesson.lesson_data.updated_at,
-      copyrightContent: copyright,
-    };
+      lessonSlug,
+      browse_data as SyntheticUnitvariantLessons[],
+      is_legacy,
+    );
 
-    return lessonDownloadsSchema.parse({
-      ...pageData,
-      nextLessons,
-      isLegacy: false,
-      isSpecialist: false,
-    });
+    const lessonDownloads = constructLessonDownloads(
+      downloads,
+      lessonSlug,
+      browse_data as SyntheticUnitvariantLessons[],
+      expired,
+    );
+
+    if (canonicalLesson) {
+      return lessonDownloadsCanonicalSchema.parse(
+        canonicalLessonDownloads,
+      ) as T;
+    } else {
+      return lessonDownloadsSchema.parse({
+        ...lessonDownloads,
+        isLegacy: false,
+        isSpecialist: false,
+      }) as T;
+    }
   };
 
+export type LessonDownloadsQuery = ReturnType<typeof lessonDownloadsQuery>;
 export default lessonDownloadsQuery;
