+++ src/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloadsCanonical.schema.ts
