+++ src/node-lib/curriculum-api-2023/queries/lessonOverview/lessonOverview.query.test.ts
@@ -182,42 +182,42 @@ describe("lessonOverview()", () => {
       ]);
     });
 
-    // it("should return an array with empty copyrightInfo if not provided", () => {
-    //   const content = [
-    //     { otherInfo: "Some other info" },
-    //     { otherInfo: "More info" },
-    //   ];
-    //   expect(getCopyrightContent(content)).toEqual([
-    //     { copyrightInfo: "" },
-    //     { copyrightInfo: "" },
-    //   ]);
-    // });
+    it("should return an array with empty copyrightInfo if not provided", () => {
+      const content = [
+        { otherInfo: "Some other info", copyrightInfo: "" },
+        { otherInfo: "More info", copyrightInfo: "" },
+      ];
+      expect(getCopyrightContent(content)).toEqual([
+        { copyrightInfo: "" },
+        { copyrightInfo: "" },
+      ]);
+    });
 
-    // it("should handle mixed content properly", () => {
-    //   const content = [
-    //     { copyrightInfo: "Copyright 2024 Example Corp." },
-    //     { otherInfo: "Some other info" },
-    //     { copyrightInfo: "Copyright 2023 Another Corp." },
-    //     { otherInfo: "More info" },
-    //   ];
-    //   expect(getCopyrightContent(content)).toEqual([
-    //     { copyrightInfo: "Copyright 2024 Example Corp." },
-    //     { copyrightInfo: "" },
-    //     { copyrightInfo: "Copyright 2023 Another Corp." },
-    //     { copyrightInfo: "" },
-    //   ]);
-    // });
+    it("should handle mixed content properly", () => {
+      const content = [
+        { copyrightInfo: "Copyright 2024 Example Corp." },
+        { otherInfo: "Some other info" },
+        { copyrightInfo: "Copyright 2023 Another Corp." },
+        { otherInfo: "More info" },
+      ];
+      expect(getCopyrightContent(content)).toEqual([
+        { copyrightInfo: "Copyright 2024 Example Corp." },
+        { copyrightInfo: "" },
+        { copyrightInfo: "Copyright 2023 Another Corp." },
+        { copyrightInfo: "" },
+      ]);
+    });
 
-    // it("should handle undefined copyrightInfo fields", () => {
-    //   const content = [
-    //     { copyrightInfo: undefined },
-    //     { otherInfo: "More info" },
-    //   ];
-    //   expect(getCopyrightContent(content)).toEqual([
-    //     { copyrightInfo: "" },
-    //     { copyrightInfo: "" },
-    //   ]);
-    // });
+    it("should handle undefined copyrightInfo fields", () => {
+      const content = [
+        { copyrightInfo: undefined },
+        { otherInfo: "More info" },
+      ];
+      expect(getCopyrightContent(content)).toEqual([
+        { copyrightInfo: "" },
+        { copyrightInfo: "" },
+      ]);
+    });
   });
   describe("getContentGuidance", () => {
     it("should return null if content.contentGuidance is null", () => {
